TWI-38 Log in and out I added a check that comes from user and added file "user_login_logout" and moved functions login and logout in it. In "helpers" added functions both check email and password.
